Moved serializer into its own file
